Add v3 verification for payloads containing issuers
This adds the ability for the gem to also verify token payloads which
contain an issuer key (iss).

Current requests in the Form Builder platform use a separate
service_slug which is sent in the params. This service_slug is used as
the key when making the request to the service token cache.

Going forward we will be sending the tokens in the Authorization: Bearer
header and the iss will be used as the key for the service token cache
to retrieve the public key.

This reduces the need for the service slug to be sent in the params
going forward as well as fit the more standardised version of a bearer
token.

The downside is that we have to call JWT.decode twice, first without
verification in order to retrieve the iss, and then again to verify the
signature once we have the public key for that requesting service.
However we felt that having the JWT carry all it needs to be verified is
the neater solution, at least for the client applications.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Matt Tei <matt.tei@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>