Retry failed validation with ingore_cache query param
If we ever need to rotate any public keys for any services signing JWT
tokens we want to do so with zero downtime. The service token cache can
receive an ignore_cache parameter which will make it look for a public
key in Kubernetes instead of the Redis cache.

Therefore we can rotate a public key in the config maps for any
application on the platform or services namespaces and if the first
validation fails the gem will make another request which should return
the public key in Kubernetes for the requested application.