Supply issuer to service token client
The issuer plays a pivotal role in the JWT authentication mechanism. It
is used when the service token cache makes a call to either the redis
key that holds the public key or more importantly when it attempts to
retreive the public key from kubernetes.

It is actually searching for a config map which contains that public
key. The name of that config map takes the form
`fb-<parameterised-service-name>-config-map` where the
parametersed-service-name can be whatever the service owner calls their
form. e.g my-amazing-form, apply-financial-deputy, leavers etc.

The actual name of this parameter changes depending on which app you are
in on the platform. Here it is known as issuer and is set as the value
to the `iss` property in the JWT.

The Editor uses a parameterised version of the service name in order to
create the config map that contains the public key for a given service
when it publishes said service. However it calls this creation the
`service_slug`.

See:

https://github.com/ministryofjustice/fb-editor/blob/201e91e249d6a61caee60a2ef0eecf7cf626b5fd/app/services/publisher/service_provisioner.rb#L28

and

https://github.com/ministryofjustice/fb-editor/blob/201e91e249d6a61caee60a2ef0eecf7cf626b5fd/app/services/publisher/service_provisioner.rb#L47

The v3 route for the service token cache then expects a path paramter
which is that service slug, but it calls the parameter `application`.

Internally the service token cache then passes the `application` as a
`service_slug` parameter on the PublicKeyService class before finally
being converted for the very last time into what the Kubernetes Adapter
thinks of as a `secret_name`. Except it's actually not a secret name at
all, it's the name of the config map that the Editor created waaaaaaay
back in this commit message.

The important thing to remember is the name of your service is what you
need to pass into the ServiceTokenClient. E.g `apply-financial-deputy`.
Eventually what actually ends up getting requested in Cloud Platform is
`fb-apply-financial-deputy-config-map`. The runner for a given service
is the thing that will have to inject the issuer when generating the JWT
therefore it will need to know its' own service slug from either the
metadata or from an injected environment variable at publishing time.

Amazing.